Ad traceroute
I still can't display these routes. An issue
is that I can't add traces to plotly after
a call show() on the figure.

Another issue is that even if I create an
entirely new figure with the new traces, I see
them in the legend but not on the map itself.
I should investigate why this is.

Another thing of not is that traceroute does
not return a list of ip addresses.